split out feature flags and proof into their own files
